Automatic extraction for 120000 samplig sites
The step of automated data extraction based on Python 2.7 and ArcGIS 10.5 was one of the key steps to implement automatic batch sampling. To accurately study the interac-tions of multi-scale factors on landscape metrics, we needed  big data from sampling sites to accurately capture the fine structural characteristics of com-plex and heterogeneous landscapes. We set 120 parallel processing programs  based on the gradients of three scales (spatial extent, spatial resolution, and classification) and selected 1000 random sampling sites for each processing program . Among them, the spatial extent was clipped to a circular shape centered on random sampling sites with a radius of 5 km, 10 km, 20 km, 30 km, 40 km, and 50 km. Then, we accessed a site’s dataset containing 120,000 sampling maps after the program of automated data ex-traction was applied to all 120 parallel processing programs with different extent, resolu-tion, and classification gradients in this study. Finally, after the batch calculation, we ac-cessed a big training dataset for multi-scale interaction analysis for 12 landscape pattern metrics on the landscape level.